Completes fragment_rate_view. Refactor file names.
