+++ app/src/main/java/com/example/animatedrating/fragment/RatingsFragment.kt
@@ -1,0 +1,44 @@
+package com.example.animatedrating.fragment
+
+import android.content.Context
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.animatedrating.R
+
+import com.example.animatedrating.fragment.dummy.DummyContent
+import com.example.animatedrating.fragment.dummy.DummyContent.DummyItem
+
+class RatingsFragment : Fragment() {
+
+
+    private var listener: OnListFragmentInteractionListener? = null
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_ratings_list, container, false)
+    }
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+        if (context is OnListFragmentInteractionListener) {
+            listener = context
+        } else {
+            throw RuntimeException(context.toString() + " must implement OnListFragmentInteractionListener")
+        }
+    }
+
+    override fun onDetach() {
+        super.onDetach()
+        listener = null
+    }
+
+    interface OnListFragmentInteractionListener {
+        // TODO: Update argument type and name
+        fun onListFragmentInteraction(item: DummyItem?)
+    }
+}
