+++ app/src/main/java/com/example/animatedrating/model/Rating.kt
@@ -1,0 +1,8 @@
+package com.example.animatedrating.model
+
+import java.io.Serializable
+
+class Rating(var name: String,var rating: Int): Serializable{
+
+    var ratingsArray = mutableListOf<Rating>()
+}
