+++ app/src/main/res/layout/fragment_rate_view.xml
@@ -1,0 +1,120 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".fragment.RateView">
+
+    <androidx.cardview.widget.CardView
+        android:layout_width="220dp"
+        android:layout_height="100dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent">
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="220dp"
+            android:layout_height="100dp"
+            android:background="@color/cardview_light_background"
+            android:clickable="true"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent">
+
+            <TextView
+                android:id="@+id/rating_text_view"
+                android:layout_margin="4dp"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="Rating:"
+                android:textSize="14sp"
+                app:layout_constraintLeft_toLeftOf="parent"
+                app:layout_constraintTop_toTopOf="parent" />
+
+            <ImageView
+                android:id="@+id/image_view1"
+                android:layout_margin="4dp"
+                android:layout_width="30dp"
+                android:layout_height="25dp"
+                android:scaleType="fitCenter"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintLeft_toRightOf="@+id/rating_text_view"/>
+
+            <ImageView
+                android:id="@+id/image_view2"
+                android:layout_margin="4dp"
+                android:layout_width="30dp"
+                android:layout_height="25dp"
+                android:scaleType="fitCenter"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintLeft_toRightOf="@+id/image_view1"/>
+
+            <ImageView
+                android:id="@+id/image_view3"
+                android:layout_margin="4dp"
+                android:layout_width="30dp"
+                android:layout_height="25dp"
+                android:scaleType="fitCenter"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintLeft_toRightOf="@+id/image_view2"/>
+
+            <ImageView
+                android:id="@+id/image_view4"
+                android:layout_margin="4dp"
+                android:layout_width="30dp"
+                android:layout_height="25dp"
+                android:scaleType="fitCenter"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintLeft_toRightOf="@+id/image_view3"/>
+
+            <ImageView
+                android:id="@+id/image_view5"
+                android:layout_margin="4dp"
+                android:layout_width="30dp"
+                android:layout_height="25dp"
+                android:scaleType="fitCenter"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintLeft_toRightOf="@+id/image_view4"/>
+
+            <EditText
+                android:id="@+id/edit_text"
+                android:layout_width="150dp"
+                android:layout_height="36dp"
+                android:hint="@string/name"
+                android:textSize="12sp"
+                android:textAlignment="center"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent" />
+
+            <Button
+                android:id="@+id/btn_save"
+                android:layout_width="80dp"
+                android:layout_height="30dp"
+                android:text="Save"
+                android:textSize="10sp"
+                app:layout_constraintRight_toLeftOf="@+id/btn_remove"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintLeft_toLeftOf="parent" />
+
+            <Button
+                android:id="@+id/btn_remove"
+                android:layout_width="80dp"
+                android:layout_height="30dp"
+                android:text="Remove"
+                android:textSize="10sp"
+                app:layout_constraintLeft_toRightOf="@+id/btn_save"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintRight_toRightOf="parent" />
+
+
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
+
+    </androidx.cardview.widget.CardView>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
