+++ app/src/main/res/layout/fragment_ratings.xml
