+++ app/src/main/res/layout/fragment_ratings_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".fragment.ratingFragment"
-    tools:listitem="@layout/fragment_rating" />
+    tools:context=".fragment.RatingsFragment"
+    tools:listitem="@layout/fragment_ratings" />
